+++ src/provider/utils/conditions/helpers.ts
@@ -15,15 +15,28 @@ export function buildConditionExpression(conditions: ItemExpression<any>[]): str
   return buildExpression(conditions, CONDITION_PREFIX);
 }
 
+function getConditionProps(conditions: ItemExpression<any>[]): string[] {
+  return conditions
+    .map(({ property, nested }) => [property, ...(nested?.length ? getConditionProps(nested) : [])])
+    .flat();
+}
+
 export function getConditionExpressionNames(conditions: ItemExpression<any>[]): AnyObject {
   return getExpressionNames(
-    conditions.map(({ property }) => property),
+    //
+    getConditionProps(conditions),
     CONDITION_PREFIX,
   );
 }
 
+function flatConditions(conditions: ItemExpression<any>[]): ItemExpression<any>[] {
+  return conditions
+    .map(({ nested, ...cond }) => [cond, ...(nested?.length ? flatConditions(nested) : [])])
+    .flat();
+}
+
 export function getConditionExpressionValues(conditions: ItemExpression<any>[]): AnyObject {
-  return getExpressionValues(conditions, CONDITION_PREFIX);
+  return getExpressionValues(flatConditions(conditions), CONDITION_PREFIX);
 }
 
 export function getConditionParams(conditions?: ItemExpression<any>[]): DBConditionParams {