+++ changelog.md
@@ -33,6 +33,7 @@ table.schema.from(entity).query.noParam({ value: 'Something' })
 - *Fix*: Transaction size validation reference after null checks
 - *Fix*: `schema.from(xxx).delete()` params no longer required if entity has no key params
 - *Fix*: `paginationToken` on `QueryResult` was marked as required.
+- *Fix*: `nested` conditions were not properly generating their `ExpressionAttributeNames` and `ExpressionAttributeValues`
 - *Fix*: `propertiesToGet` deprecated on `list`/`listAll` methods, as `propertiesToRetrieve` is used on `get`/`batchGet` and is more descriptive. Will be removed on v3.
 
 - *Testing*: Type checks enabled and type tests incorporated on every test file. Previously, updates to types could easily break usages via type-check breaks. Actual runtime results were OK but one of the huge benefits of our lib were basically vulnerable to badly tested modifications